+++ packages/monorepo/src/spawn.ts
@@ -1,7 +1,0 @@
-import { spawnSync as childProcessSpawn } from 'child_process';
-
-export const spawn = (command: string, args: string[] = []) => {
-  return childProcessSpawn(command, args, {
-    stdio: 'inherit',
-  });
-};