+++ packages/react-auth-strapi/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.59](https://github.com/ttoss/ttoss/compare/@ttoss/react-auth-strapi@0.2.58...@ttoss/react-auth-strapi@0.2.59) (2026-01-07)
+
+**Note:** Version bump only for package @ttoss/react-auth-strapi
+
 ## [0.2.58](https://github.com/ttoss/ttoss/compare/@ttoss/react-auth-strapi@0.2.57...@ttoss/react-auth-strapi@0.2.58) (2026-01-01)
 
 **Note:** Version bump only for package @ttoss/react-auth-strapi